[Mips] Fix enumeral and non-enumeral type in conditional expression warning
No functional changes.

llvm-svn: 227297

+++ lld/lib/ReaderWriter/ELF/Mips/MipsELFFlagsMerger.cpp
@@ -71,7 +71,7 @@ std::error_code MipsELFFlagsMerger::merge(uint8_t newClass, uint32_t newFlags) {
   // We support two ABI: O32 and N64. The last one does not have
   // the corresponding ELF flag.
   uint32_t inAbi = newFlags & EF_MIPS_ABI;
-  uint32_t supportedAbi = _is64Bit ? 0 : EF_MIPS_ABI_O32;
+  uint32_t supportedAbi = _is64Bit ? 0 : uint32_t(EF_MIPS_ABI_O32);
   if (inAbi != supportedAbi)
     return make_dynamic_error_code(Twine("Unsupported ABI"));
